Debug task_struct : user space now running

+++ src/lesson06/include/sched.h
@@ -58,7 +58,7 @@ struct mm_struct {
 struct task_struct {
 	struct cpu_context cpu_context;
 	long state;
-	int blocked_by;
+	unsigned long blocked_by;
 	long counter;
 	long priority;
 	long preempt_count;
@@ -77,7 +77,7 @@ extern void exit_process(void);
 
 #define INIT_TASK \
 /*cpu_context*/ { { 0,0,0,0,0,0,0,0,0,0,0,0,0}, \
-/* state etc */	 0,0,15, 0, PF_KTHREAD, \
+/* state etc */	 0,0,15, 0, -1,  PF_KTHREAD, \
 /* mm */ { 0, 0, {{0}}, 0, {0}} \
 }
 #endif
